+++ cmake/ci_toolchain.cmake
@@ -1,2 +1,0 @@
-set(CMAKE_CXX_COMPILER "/usr/bin/clang++-19")
-set(CMAKE_C_COMPILER "/usr/bin/clang-19")