+++ src/text_io.cpp
@@ -15,18 +15,14 @@
 #include "view.h"
 
 [[nodiscard]] opt_lines_t open_file(const std::string& file) {
-    if (!get_file_size(file)) {
-        return {};
-    }
+    if (!get_file_size(file)) { return {}; }
 
     auto ret = std::vector<std::string>();
     std::string line = "";
     char ch;
 
     std::ifstream ifs(file);
-    if (ifs.fail()) {
-        return {};
-    }
+    if (ifs.fail()) { return {}; }
 
     while (ifs.get(ch)) {
         switch (ch) {
@@ -46,9 +42,7 @@
 
     // If there's no newlines in the stream at all, it never gets added to
     // the vector
-    if (line.size()) {
-        ret.push_back(line);
-    }
+    if (line.size()) { ret.push_back(line); }
 
     return ret;
 }
@@ -59,15 +53,14 @@
 
     try {
         return uint_t(fs::file_size(p));
-    } catch (const fs::filesystem_error&) {
-        return 0;
-    }
+    } catch (const fs::filesystem_error&) { return 0; }
 }
 
-[[nodiscard]] WriteData write_to_file(Model* model) {
-    if (model->filename == "") {
-        return WriteData();
-    }
+[[nodiscard]] WriteData write_to_file(Model* model, std::optional<std::string> filename_input) {
+    if (!filename_input.has_value() && model->filename == "") { return WriteData(); }
+
+    if (filename_input.has_value()) { model->filename = filename_input.value(); }
+
     std::ofstream out(model->filename);
     for (auto&& line : model->buf) {
         rtrim(line);
@@ -89,20 +82,20 @@ void rtrim(std::string& str) {
     std::string line;
 
     while (std::getline(ss, line)) {
-        if (line.back() == '\r') {
-            line.pop_back();
-        }
+        if (line.back() == '\r') { line.pop_back(); }
         result.push_back(rawterm::raw_str(line));
     }
 
     return result;
 }
 
+// TODO: Investigate if symbols should count here
 [[nodiscard]] bool is_letter(const char& c) {
-    std::array<char, 52> alphabet = {
-        'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r',
-        's', 't', 'u', 'v', 'w', 'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J',
-        'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'};
+    std::array<char, 62> alphabet = {
+        'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p',
+        'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F',
+        'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V',
+        'W', 'X', 'Y', 'Z', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9'};
 
     return std::find(alphabet.begin(), alphabet.end(), c) != alphabet.end();
 }
@@ -187,9 +180,7 @@ void rtrim(std::string& str) {
         }
 
         auto logger = spdlog::get("basic_logger");
-        if (logger != nullptr) {
-            logger->info("Shell cmd run: `" + cmd + "`");
-        }
+        if (logger != nullptr) { logger->info("Shell cmd run: `" + cmd + "`"); }
 
         if (!resp.out.empty() && resp.out.at(resp.out.size() - 1) == '\n') {
             resp.out = resp.out.substr(0, resp.out.size() - 1);
@@ -213,8 +204,6 @@ void rtrim(std::string& str) {
 
 [[nodiscard]] int first_non_whitespace(const std::string& line) {
     std::size_t ret = line.find_first_not_of(WHITESPACE, 0);
-    if (ret == std::string::npos) {
-        return -1;
-    }
+    if (ret == std::string::npos) { return -1; }
     return int32_t(ret);
 }