+++ src/view.cpp
@@ -25,6 +25,7 @@
 View::View(Controller* controller, const rawterm::Pos dims)
     : ctrlr_ptr(controller), view_size(dims), cur(rawterm::Cursor()) {
     view_models.reserve(8);
+    prev_tab_bar.reserve(uint_t(dims.horizontal));
 }
 
 void View::add_model(Model* m) {
@@ -61,9 +62,7 @@ void View::draw_screen() {
         }
     }
 
-    if (visible_tab_bar() && cur.vertical == 1) {
-        cur.move_down();
-    }
+    if (visible_tab_bar() && cur.vertical == 1) { cur.move_down(); }
 
     if (get_active_model()->type == ModelType::META) {
         rawterm::Cursor::cursor_hide();
@@ -76,9 +75,7 @@ void View::draw_screen() {
     std::string screen;
 
     // Draw tab bar
-    if (view_models.size() > 1) {
-        screen += render_tab_bar();
-    }
+    if (view_models.size() > 1) { screen += render_tab_bar(); }
 
     // Get displayable subrange
     std::size_t end = std::min(std::size_t(view_size.vertical - 2), get_active_model()->buf.size());
@@ -91,12 +88,10 @@ void View::draw_screen() {
 
     const std::size_t start_idx = get_active_model()->view_offset + 1;
     const std::size_t vert_offset = get_active_model()->vertical_offset;
-    for (const auto [idx, line] : enumerate<std::string>(viewable_range, start_idx)) {
+    for (const auto& [idx, line] : enumerate<std::string>(viewable_range, start_idx)) {
         if (LINE_NUMBERS) {
             rawterm::Color c = COLOR_UI_BG;
-            if (idx == get_active_model()->current_line + 1) {
-                c = COLOR_DARK_YELLOW;
-            }
+            if (idx == get_active_model()->current_line + 1) { c = COLOR_DARK_YELLOW; }
             screen +=
                 rawterm::set_foreground(std::format("{:>{}}\u2502", idx, line_number_offset), c);
         }
@@ -106,17 +101,13 @@ void View::draw_screen() {
             continue;
         }
 
-        if (vert_offset) {
-            screen += "\u00AB";
-        }
+        if (vert_offset) { screen += "\u00AB"; }
 
         if (line.size() > viewable_hor_len) {
             const std::size_t draw_len = viewable_hor_len - 2 - (vert_offset ? 1 : 0);
             screen += line.substr(vert_offset, draw_len);
 
-            if (line.size() > draw_len + vert_offset + 3) {
-                screen += "\u00BB";
-            }
+            if (line.size() > draw_len + vert_offset + 3) { screen += "\u00BB"; }
         } else {
             screen += line.substr(vert_offset, line.size());
         };
@@ -143,32 +134,33 @@ void View::draw_screen() {
 }
 
 void View::draw_tab_bar() {
-    if (view_models.size() == 1) {
-        return;
-    }
+    if (view_models.size() == 1) { return; }
+
+    std::string new_tab_bar = render_tab_bar();
+    if (new_tab_bar == prev_tab_bar) { return; }
 
     rawterm::Pos starting_cur_pos = cur;
     cur.move(1, 1);
     rawterm::clear_line();
-    std::print("{}", render_tab_bar());
+    std::print("{}", new_tab_bar);
+    prev_tab_bar = new_tab_bar;
     cur.move(starting_cur_pos);
 }
 
 const std::string View::render_tab_bar() const {
     // Extract base filenames
     std::vector<std::string> base_filenames;
+    std::size_t filename_max_length = 0;
+
     for (const auto& model : view_models) {
         std::string full_path = model->filename;
         std::size_t last_slash = full_path.find_last_of("/\\");
         std::string filename =
             (last_slash != std::string::npos) ? full_path.substr(last_slash + 1) : full_path;
         base_filenames.push_back(filename);
-    }
 
-    // Find longest filename
-    std::size_t filename_max_length = 0;
-    for (const auto& name : base_filenames) {
-        filename_max_length = std::max(filename_max_length, name.size());
+        // Find longest filename
+        filename_max_length = std::max(filename_max_length, filename.size());
     }
 
     std::string ret;
@@ -183,9 +175,7 @@ const std::string View::render_tab_bar() const {
                 display_name = display_name.substr(0, filename_max_length);
             }
 
-            if (view_models.at(i)->unsaved) {
-                display_name += "*";
-            }
+            if (view_models.at(i)->unsaved) { display_name += "*"; }
 
             if (i == active_model) {
                 ret += rawterm::inverse(display_name);
@@ -244,22 +234,16 @@ void View::draw_line(const Draw_Line_dir::values redraw_prev) {
     }
 
     const std::size_t vert_offset = get_active_model()->vertical_offset;
-    if (curr_line.empty() || (vert_offset && curr_line.size() < vert_offset)) {
-        return line;
-    }
+    if (curr_line.empty() || (vert_offset && curr_line.size() < vert_offset)) { return line; }
 
     // Truncate
     if (curr_line.size() > viewable_hor_len) {
-        if (vert_offset) {
-            line += "\u00AB";
-        }
+        if (vert_offset) { line += "\u00AB"; }
 
         const std::size_t draw_len = viewable_hor_len - 2 - (vert_offset ? 1 : 0);
         line += curr_line.substr(vert_offset, draw_len);
 
-        if (curr_line.size() > draw_len + vert_offset + 2) {
-            line += "\u00BB";
-        }
+        if (curr_line.size() > draw_len + vert_offset + 2) { line += "\u00BB"; }
 
     } else {
         line += curr_line.substr(vert_offset, curr_line.size());
@@ -307,8 +291,7 @@ const std::string View::render_status_bar() const {
         right += "| [" + std::to_string(ctrlr_ptr->models.size()) + "] ";
     }
 
-    const std::string cursor_pos = "| " + std::to_string(get_active_model()->current_line + 1) +
-                                   ":" + std::to_string(get_active_model()->current_char + 1) + " ";
+    const std::string cursor_pos = "|" + render_cursor_coords() + " ";
 
     right += cursor_pos;
 
@@ -328,7 +311,7 @@ const std::string View::render_status_bar() const {
     /// Calculate whitespace to insert
     /// https://excalidraw.com/#json=_ysE8iOIfRygS1qx7sdfT,8MWhqK-DAdUf__mxefZp8w
     const std::size_t available_center_space =
-        std::size_t(view_size.horizontal) - left.size() - right.size();
+        std::size_t(view_size.horizontal) - left.size() - rawterm::raw_size(right);
     const std::size_t filename_start_offset =
         (available_center_space - visible_filename.size()) / 2;
     const std::size_t filename_start = left.size() + filename_start_offset;
@@ -339,7 +322,7 @@ const std::string View::render_status_bar() const {
     std::string ret = left + std::string(std::size_t(lhs_space), ' ');
     ret += visible_filename + std::string(std::size_t(rhs_space), ' ') + right;
 
-    assert(ret.size() == static_cast<std::size_t>(view_size.horizontal));
+    assert(rawterm::raw_size(ret) == static_cast<std::size_t>(view_size.horizontal));
     return rawterm::set_background(ret, COLOR_UI_BG);
 }
 
@@ -358,9 +341,7 @@ void View::display_message(std::string msg, std::optional<rawterm::Color> color)
     cur.move({view_size.vertical, 1});
     rawterm::clear_line();
 
-    if (color.has_value()) {
-        msg = rawterm::set_foreground(msg, color.value());
-    }
+    if (color.has_value()) { msg = rawterm::set_foreground(msg, color.value()); }
 
     std::print("{}", msg);
     cur.move(prev_pos);
@@ -372,9 +353,7 @@ void View::display_message(std::string msg, std::optional<rawterm::Color> color)
 // crashes that were down to the offset being wrong after editing text (I think)
 [[nodiscard]] std::size_t View::clamp_horizontal_movement(const int offset) {
     const int line_pos = static_cast<int>(get_active_model()->current_line) + offset;
-    if (line_pos < 0 || line_pos > int32_t(get_active_model()->buf.size())) {
-        return 0;
-    }
+    if (line_pos < 0 || line_pos > int32_t(get_active_model()->buf.size())) { return 0; }
 
     std::string_view line_moving_to =
         get_active_model()->buf.at(static_cast<std::size_t>(line_pos));
@@ -391,9 +370,7 @@ void View::display_message(std::string msg, std::optional<rawterm::Color> color)
     if (get_active_model()->vertical_offset &&
         uint_t(cur.horizontal) == (LINE_NUMBERS ? line_number_offset + 3 : 0)) {
         get_active_model()->current_char--;
-        if (get_active_model()->vertical_offset == 2) {
-            get_active_model()->vertical_offset--;
-        }
+        if (get_active_model()->vertical_offset == 2) { get_active_model()->vertical_offset--; }
         get_active_model()->vertical_offset--;
         return true;
     } else if (get_active_model()->current_char) {
@@ -407,9 +384,7 @@ void View::display_message(std::string msg, std::optional<rawterm::Color> color)
 
 [[maybe_unused]] bool View::cursor_up(unsigned int count) {
     // If no line above, do nothing
-    if (!(get_active_model()->current_line)) {
-        return false;
-    }
+    if (!(get_active_model()->current_line)) { return false; }
 
     std::size_t horizontal_clamp = clamp_horizontal_movement(int32_t(-count));
     get_active_model()->current_line -= count;
@@ -437,9 +412,7 @@ void View::display_message(std::string msg, std::optional<rawterm::Color> color)
 
 [[maybe_unused]] bool View::cursor_down(unsigned int count) {
     // If we're on the last line, do nothing
-    if (get_active_model()->current_line >= get_active_model()->buf.size() - 1) {
-        return false;
-    }
+    if (get_active_model()->current_line >= get_active_model()->buf.size() - 1) { return false; }
 
     std::size_t horizontal_clamp = clamp_horizontal_movement(int32_t(count));
     get_active_model()->current_line += count;
@@ -473,9 +446,7 @@ void View::display_message(std::string msg, std::optional<rawterm::Color> color)
     // Only scroll if we're still in the line
     std::string_view curr_line = get_active_model()->buf.at(get_active_model()->current_line);
     const std::size_t line_size = curr_line.size();
-    if (get_active_model()->current_char == line_size) {
-        return false;
-    }
+    if (get_active_model()->current_char == line_size) { return false; }
 
     if (cur.horizontal < view_size.horizontal - 2) {
         get_active_model()->current_char++;
@@ -483,9 +454,7 @@ void View::display_message(std::string msg, std::optional<rawterm::Color> color)
         return false;
     } else {
         get_active_model()->current_char++;
-        if (!get_active_model()->vertical_offset) {
-            get_active_model()->vertical_offset++;
-        }
+        if (!get_active_model()->vertical_offset) { get_active_model()->vertical_offset++; }
         get_active_model()->vertical_offset++;
         return true;
     }
@@ -520,13 +489,11 @@ void View::center_current_line() {
 }
 
 void View::set_current_line(const unsigned int lineno) {
-    if (lineno > get_active_model()->buf.size()) {
-        return;
-    }
+    if (lineno > get_active_model()->buf.size()) { return; }
 
     get_active_model()->current_line = lineno - 1;
     uint_t half_view = static_cast<uint_t>(std::floor(view_size.vertical / 2));
-    uint_t curr_char = get_active_model()->current_char + line_number_offset + 2;
+    std::size_t curr_char = get_active_model()->current_char + line_number_offset + 2;
     if (get_active_model()->vertical_offset) {
         curr_char -= (get_active_model()->vertical_offset - 1);
     }
@@ -542,9 +509,7 @@ void View::set_current_line(const unsigned int lineno) {
 
 void View::get_git_branch() {
     auto resp = shell_exec("git rev-parse --abbrev-ref HEAD");
-    if (resp.has_value()) {
-        git_branch = resp.value().out;
-    }
+    if (resp.has_value()) { git_branch = resp.value().out; }
 }
 
 void View::tab_new() {
@@ -591,7 +556,7 @@ void View::tab_prev() {
 void View::change_model_cursor() {
     const uint_t vertical =
         get_active_model()->current_line - get_active_model()->view_offset + visible_tab_bar() + 1;
-    uint_t horizontal = get_active_model()->current_char + uint_t(line_number_offset) + 2;
+    std::size_t horizontal = get_active_model()->current_char + uint_t(line_number_offset) + 2;
     if (get_active_model()->vertical_offset) {
         horizontal -= get_active_model()->vertical_offset - 1;
     }
@@ -600,9 +565,7 @@ void View::change_model_cursor() {
 }
 
 bool View::set_buffer(const std::size_t bufnr, const std::size_t model_len) {
-    if (bufnr >= model_len) {
-        return false;
-    }
+    if (bufnr >= model_len) { return false; }
 
     view_models.at(active_model) = &ctrlr_ptr->models.at(std::size_t(bufnr));
     change_model_cursor();
@@ -618,4 +581,29 @@ void View::draw_overlay(std::span<std::string> contents, std::string_view title)
 
     auto border = rawterm::Border(region).set_padding(1).set_title(std::format(" {} ", title));
     border.draw(cur, contents);
+}
+
+[[nodiscard]] std::string View::render_cursor_coords() const {
+    std::string ret = "";
+    const std::size_t curr_char = get_active_model()->current_char + 1;
+
+    if (curr_char >= LINE_BORDER) { ret += COLOR_ALERT; }
+
+    ret += " ";
+
+    ret += std::to_string(get_active_model()->current_line + 1);
+    ret += ":";
+    ret += std::to_string(curr_char);
+
+    if (curr_char >= LINE_BORDER) {
+        // We need to re-open the base colour for the statusbar
+
+        std::string colors = std::to_string(COLOR_UI_BG.red) + ";";
+        colors += std::to_string(COLOR_UI_BG.blue) + ";";
+        colors += std::to_string(COLOR_UI_BG.green) + "m";
+
+        ret += "\x1b[" + colors;
+    }
+
+    return ret;
 }