+++ tests/view_test.cpp
@@ -28,9 +28,7 @@ TEST_CASE("add_model", "[view]") {
     REQUIRE(v.active_model == 0);
     REQUIRE(v.get_active_model()->filename == "test_file.txt");
 
-    if (LINE_NUMBERS) {
-        REQUIRE(v.line_number_offset == 2);
-    }
+    if (LINE_NUMBERS) { REQUIRE(v.line_number_offset == 2); }
 }
 
 TEST_CASE("get_active_model", "[view]") {
@@ -606,4 +604,19 @@ TEST_CASE("set_buffer", "[view]") {
     c.view.set_buffer(0, c.models.size());
     REQUIRE(c.view.view_models.at(0)->filename == "tests/fixture/lorem_ipsum.txt");
     REQUIRE(c.view.get_active_model()->current_line == 0);
+}
+
+TEST_CASE("render_cursor_coords", "[view]") {
+    Controller c;
+    c.create_view("tests/fixture/very_long_line.txt", 0);
+    c.view.cursor_end_of_line();
+
+    const std::string statusbar = c.view.render_cursor_coords();
+
+    REQUIRE(statusbar.contains(COLOR_ALERT));
+
+    std::size_t first_bg = statusbar.find(COLOR_UI_BG.to_str());
+    REQUIRE(first_bg != std::string::npos);
+    std::size_t second_bg = statusbar.substr(first_bg, statusbar.size()).find(COLOR_UI_BG.to_str());
+    REQUIRE(second_bg != std::string::npos);
 }